requirements.txt
Django==5.0.6
djangorestframework==3.15.2
google-generativeai==0.5.2
python-dotenv==1.0.1
gunicorn==23.0.0
whitenoise==6.7.0